Move existing base32 tests out of test_util.py.
